--- a/hw10_program_optimization/stat
+++ b/hw10_program_optimization/stat
@@ -1,0 +1,8 @@
+name             old time/op    new time/op    delta
+GetDomainStat-4     197ms ± 1%       0ms ± 1%   -99.99%  (p=0.016 n=4+5)
+
+name             old alloc/op   new alloc/op   delta
+GetDomainStat-4     134MB ± 0%       0MB ± 0%   -99.99%  (p=0.008 n=5+5)
+
+name             old allocs/op  new allocs/op  delta
+GetDomainStat-4     1.70M ± 0%     0.00M ± 0%  -100.00%  (p=0.008 n=5+5)