--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,11 +1,9 @@
 package hw10programoptimization
 
 import (
+	"bufio"
 	"encoding/json"
-	"fmt"
 	"io"
-	"io/ioutil"
-	"regexp"
 	"strings"
 )
 
@@ -22,46 +20,28 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
+	scanner := bufio.NewScanner(r)
+	scanner.Split(bufio.ScanLines)
 
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
+	user := new(User)
 	result := make(DomainStat)
+	suffix := "." + domain
 
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
+	// read line by line
+	for scanner.Scan() {
+		if !strings.Contains(scanner.Text(), domain) {
+			continue
 		}
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		if err := json.Unmarshal(scanner.Bytes(), &user); err != nil {
+			return result, err
 		}
+
+		if strings.HasSuffix(user.Email, suffix) {
+			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]++
+		}
+		user = &User{}
 	}
+
 	return result, nil
 }