--- a/hw10_program_optimization/stats_optimization_test.go
+++ b/hw10_program_optimization/stats_optimization_test.go
@@ -1,3 +1,4 @@
+//go:build bench
 // +build bench
 
 package hw10programoptimization
@@ -13,8 +14,7 @@ import (
 const (
 	mb          uint64 = 1 << 20
 	memoryLimit uint64 = 30 * mb
-
-	timeLimit = 300 * time.Millisecond
+	timeLimit          = 300 * time.Millisecond
 )
 
 // go test -v -count=1 -timeout=30s -tags bench .